Fix test & print seed for failed tests
Closes #38
Closes #39
