+++ src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,2 @@
-mr.cookie.matrix.listeners.JUnitAugmentationListener
+mr.cookie.matrix.listeners.JUnitRandomSeedListener
+mr.cookie.matrix.listeners.JUnitElapsedTimeListener
